+++ cmd/build.go
@@ -2,29 +2,22 @@ package cmd
 
 import (
 	"fmt"
-	"forge/internal/config"
+	"forge/internal/logger"
 )
 
-func Build() error {
-	_, err := fmt.Println("Building the project...")
-	if err != nil {
+func Build(args ...string) error {
+	if len(args) < 1 {
+		logger.Error("Preset not specified. Use forge build <preset> (debug or release).")
+		return fmt.Errorf("Failed to build project.")
+	}
+
+	preset := args[0]
+	logger.Infof("Building with CMake preset %s", preset)
+
+	if err := runCommand("cmake", "--preset", preset); err != nil {
 		return err
 	}
-
-	
-
-	preset := config.Get().Build.Type
-	
-	if preset == "" {
-		preset = "debug"
-	}
-
-	err = runCommand("cmake", "--preset", preset)
-	if err != nil {
-		return err
-	}
-	err = runCommand("cmake", "--build", "--preset", preset)
-	if err != nil {
+	if err := runCommand("cmake", "--build", "--preset", preset); err != nil {
 		return err
 	}
 	return nil